Set premis_object size column to BIGINT
Re-build SQL and models

+++ data/sql/lib.model.schema.sql
@@ -878,7 +878,7 @@ CREATE TABLE `premis_object`
 	`filename` VARCHAR(1024),
 	`last_modified` DATETIME,
 	`date_ingested` DATE,
-	`size` INTEGER,
+	`size` BIGINT,
 	`mime_type` VARCHAR(255),
 	`created_at` DATETIME  NOT NULL,
 	`updated_at` DATETIME  NOT NULL,
